Add files via upload
my first ever machine learning algorithm! This is a simple decision tree that takes the commonly utilized iris dataset and creates a model that predicts what species an iris would be (out of the three possible species classes). The first section is an exploratory analysis that creates pretty graphs. The second is our decision tree, with an output in the form of a Data Frame to easily see which observations were incorrected categorized, and a side by side comparison of our test data and our decision tree's predictions. The third section shows different methods of visualizing the decision tree. All sources used for this are listed.