Add files via upload
These shapefiles of CA state and county boundaries will be used to create a map in geopandas. The map is an extension of the chapter 2 exercise from Hands On Machine Learning (2nd edition) by Aurélien Géron (github repository: homl.info). The lat/long points are from the chapter 2 files. The shapefiles are from data.ca.gov/dataset/ca-geographic-boundaries. This will serve as an online repository for myself to easier access the files I want to use that are not within the HOML2 textbook.