Exploratory Analysis + Map Making
The first section of this program is entirely from Hands On Machine Learning (2nd edition). The second section on map making is a combination of the second chapter of HOML2 and my own code (with various sources listed). It was developed and executed in a Google's Colab cloud implementation of Jupyter Notebooks.